refactor: reorganize charge sensor routines and enhance functionality
- Moved charge sensor routines into a dedicated `charge_sensor` module for better structure and maintainability.
- Introduced new routines: `find_sensor_peak`, `run_compensation`, and `charge_sensor_csd_readout` to improve charge sensing capabilities.
- Updated imports in `__init__.py` to reflect the new module structure and included new functions in the public API.
- Enhanced documentation for new routines, detailing their functionality and usage in quantum dot measurements.
- Improved overall organization of charge sensor codebase to facilitate future development and testing.
